requirements.txt
discord.py>=2.3.2,<3.0
aiohttp>=3.9.3,<4.0
customtkinter>=5.2.0
